requirements.txt
torch==2.3.1
torchvision
numpy
pandas
scikit-learn
python-dotenv
requests
